Added menu, basic timer, texts & Roboto font
Made most of menu, and added intents to call game from specified
buttons. Added selection of Roboto fonts to use for in-game text. Also
added basic timer that simply says "Too slow!" when it finishes, need to
connect this to game logic and add leftover time to a score at the end
of each stage. Also removed Camera class as it was irrelevant.