Improved connectedness check, fixed pause/resume
Implemented a breadth-first graph search, the connectedness is then
checked by traversing along specified paths using that search algorithm.
Also implementing the parcelables was largely needless and the state is
saved by simply making sure the app exits cleanly - but implemented a
method to pause and resume the timer when exiting the app.